refactor: (#51) Merge 충돌 해결을 위한 GameHeader 재작성
